+++ lib/src/app_constants.dart
@@ -1,0 +1,5 @@
+class AppConstants {
+  AppConstants._();
+
+  static const MAX_ITEMS = 50;
+}