+++ lib/src/repo/caching/caching_repository.dart
@@ -1,0 +1,22 @@
+import 'package:moor_db_viewer/src/model/filter/filter_data.dart';
+import 'package:moor_db_viewer/src/repo/caching/caching_repo.dart';
+
+class CachingRepository extends CachingRepo {
+  static CachingRepository _instance;
+  final _filterData = Map<String, FilterData>();
+
+  CachingRepository._();
+
+  factory CachingRepository.instance() => _instance ??= CachingRepository._();
+
+  FilterData getFilterDataForTable(String entityName) {
+    if (!_filterData.containsKey(entityName)) {
+      _filterData[entityName] = FilterData();
+    }
+    return _filterData[entityName];
+  }
+
+  void updateFilterData(String entityName, FilterData filterData) {
+    _filterData[entityName] = filterData;
+  }
+}