+++ lib/src/style/theme_dimens.dart
@@ -7,6 +7,7 @@ class ThemeDimens {
   static const padding16 = 16.0;
   static const padding24 = 24.0;
   static const padding32 = 32.0;
+  static const padding48 = 48.0;
   static const padding72 = 72.0;
   static const heightFilterButton = 80.0;
 }